apps/ai/pyproject.toml
[project]
name = "pick-photo-ai-server"
version = "0.1.0"
description = "Internal AI service for Pick Photo face detection and ID-photo generation"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "opencv-python-headless>=4.9",
  "Pillow>=10.0",
  "pydantic>=2.0",
  "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
